Project4/requirements.txt
aiosignal==1.3.1
ale-py==0.8.1
asttokens==3.0.0
attrs==24.2.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
Box2D==2.3.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
executing==2.1.0
Farama-Notifications==0.0.4
ffmpeg==1.4
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
ipython==8.30.0
ipywidgets==8.1.5
jax-jumpy==1.0.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
nptyping==2.5.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==5.29.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
ray==2.39.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
shellingham==1.5.4
Shimmy==0.2.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tensorboardX==2.6.2.2
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
